pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "file-search"
version = "0.1.0"
description = "MCP server for AI-powered company document search. No API key needed."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docling>=2.55.0",
    "mcp[cli]>=1.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.9",
]
optimizations = [
    "pypdf>=4.0.0",       # Fast PDF text layer detection (skip OCR)
    "docx2txt>=0.8",      # Fast DOCX parsing (3-5x faster than Docling)
]

[project.scripts]
fs-explorer-mcp = "fs_explorer.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/fs_explorer"]
